refactor(data): the four acquisition methods and the global lock order come out
Database loses AcquireUserRow, AcquireClientRowShared, AcquireUserSessionRowById and
GetUserSessionClientsByClientId, in the interface, commondb (AcquireClientRowSharedWith and
sortedDistinctSessionIds with them), all four dialects and the regenerated mock. The global
order #297 built, users then clients shared then user_sessions then grants, existed only to
prevent engine-detected deadlocks, and a deadlock is now answered by RunInTransaction rerunning
the victim, so the order was cost without a property behind it: an extra UPDATE on every code
issued and every session started, a shared lock on every bump, and a rule every future
transaction had to remember.

Integrity never rested on it. The one ordering kept on purpose is #139's: a termination
deletes the session row first, and issuance takes that row before inserting its code, in one
transaction, so a code can never slip between a termination's sweep and its commit. That pair
and its RCSI twin keep every assertion they had.

Call sites return to their shapes before the order: issuance is the session row then the
insert; StartNewUserSession and BumpUserSession are their two writes; DeleteClient is tokens
then the client. DeleteUser keeps its explicit session loop and drops the user-row acquisition:
starting on the session rows is what lets it wait behind a termination, the replay or a
ceremony instead of deadlocking with them, and it was measured clean against issuance in both
orderings on all four engines. The replay keeps its session acquisition for the same local
reason. Every comment at a kept-order site now says why that transaction's first statement is
first and obliges no other site; no doc comment states a lock order.

Tests: the four tests/data/lock_order*_test.go files, commondb/client_delete_order_test.go,
user/usersession_manager_lock_order_test.go and handlers/lock_order_test.go are gone. The four
that pin a kept order survive next to what they test: the sweep's order in revocation_test.go,
the replay's in handler_token_test.go, the replay-versus-termination pair in a new
tests/data/replay_ordering_test.go, and the backfill's first statement where it was. The
issuance call-order pin is now a subtest of TestHandleIssueGet_TheAcquisitionOrdersTheInsert.
The blocked-party harness takes its rows with AcquireClientRow on a test client.

CLAUDE.md gains item 6 under Important Patterns: how to open a transaction here and what may
be assumed about concurrent ones.

Unit green on all three modules; data green on mysql, postgres, mssql and sqlite; integration
green on sqlite locally, four engines in the check suite. Five mutations against the kept-order
pins, all caught.

Refs #301
