+++ src/authserver/internal/handlers/handler_auth_issue_test.go
@@ -669,8 +669,6 @@ func TestHandleIssueGet(t *testing.T) {
 		// reason, which is why neither the audit nor the clear runs (#139).
 		commitError := errors.New("commit failed")
 		expectRunInTransactionThenFail(database, issuanceTx, commitError)
-		database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-		database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(true, nil).Once()
 
 		httpHelper.On("InternalServerError", rr, req, mock.MatchedBy(func(err error) bool {
@@ -725,6 +723,72 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 		}
 	}
 
+	// The transaction's shape is the whole of what the handler contributes to the #139 property,
+	// so it is pinned by sequence: one RunInTransaction, and inside it the acquisition, then the
+	// insert, with no other statement between. It cannot be seen at the data tier, which drives
+	// the statements by hand, and it cannot be seen from outside at all: the acquisition is a
+	// single-row UPDATE of a column nothing reads. A mock recording the sequence is the only
+	// place the handler's own choice is observable. No other row is taken ahead of the session
+	// row: the repository imposes no order between transactions, and a deadlock with one that
+	// takes these rows the other way round is answered by the helper rerunning this body (#301).
+	t.Run("The transaction is the acquisition then the insert, and nothing else", func(t *testing.T) {
+		httpHelper := mocks_handlerhelpers.NewHttpHelper(t)
+		authHelper := mocks_handlerhelpers.NewAuthHelper(t)
+		templateFS := &mocks_test.TestFS{}
+		codeIssuer := mocks_oauth.NewCodeIssuer(t)
+		tokenIssuer := mocks_oauth.NewTokenIssuer(t)
+		database := mocks_data.NewDatabase(t)
+		auditLogger := mocks_audit.NewAuditLogger(t)
+		userSessionManager := mocks_user.NewUserSessionManager(t)
+		permissionChecker := mocks_user.NewPermissionChecker(t)
+
+		handler := HandleIssueGet(httpHelper, authHelper, templateFS, codeIssuer, tokenIssuer, database,
+			auditLogger, userSessionManager, permissionChecker)
+
+		req := requestWithSessionIdentifier(t, liveSessionIdentifier)
+		rr := httptest.NewRecorder()
+
+		authContext := issuanceAuthContext("")
+		authHelper.On("GetAuthContext", req).Return(authContext, nil)
+		stubLiveSession(database, 123)
+
+		// Every statement on the transaction records itself, so the assertion below is about the
+		// sequence rather than about each call having happened somewhere. The stub notes the
+		// transaction's edges where the helper places them: "begin" before the body and "commit"
+		// after a nil return.
+		var order []string
+		note := func(what string) func(mock.Arguments) {
+			return func(mock.Arguments) { order = append(order, what) }
+		}
+		expectRunInTransaction(database, issuanceTx, func(edge string) { order = append(order, edge) })
+		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).
+			Run(note("session row")).Return(true, nil).Once()
+		codeIssuer.On("CreateAuthCode", issuanceTx, mock.Anything).Run(note("insert")).
+			Return(&models.Code{Id: 1, Code: "test-code", ClientId: 1,
+				RedirectURI: "https://example.com/callback", State: "test-state"}, nil)
+
+		auditLogger.On("Log", constants.AuditCreatedAuthCode, mock.Anything).Return()
+		authHelper.On("ClearAuthContext", rr, req).Return(nil)
+
+		armIssueGate(database, userSessionManager, permissionChecker, authContext.RedirectURI)
+
+		handler.ServeHTTP(rr, req)
+
+		assert.Equal(t, []string{"begin", "session row", "insert", "commit"}, order,
+			"the session row is taken before the insert, on the same transaction, and nothing else is")
+
+		// And the ordering is only worth anything if all of it is on ONE transaction: a row taken
+		// on a different connection is released the moment that statement autocommits. Both
+		// expectations above name issuanceTx, so a statement on any other handle fails the mock.
+		assert.Equal(t, http.StatusFound, rr.Code)
+		assert.Contains(t, rr.Header().Get("Location"), "code=test-code")
+
+		httpHelper.AssertExpectations(t)
+		authHelper.AssertExpectations(t)
+		database.AssertExpectations(t)
+		codeIssuer.AssertExpectations(t)
+	})
+
 	t.Run("The row is gone, so the ceremony restarts at level 1 and nothing is inserted", func(t *testing.T) {
 		httpHelper := mocks_handlerhelpers.NewHttpHelper(t)
 		authHelper := mocks_handlerhelpers.NewAuthHelper(t)
@@ -756,8 +820,6 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 				order = append(order, "rollback")
 			}
 		})
-		database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-		database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(false, nil).Once()
 
 		var savedAuthContext *oauth.AuthContext
@@ -833,8 +895,6 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 				order = append(order, "rollback")
 			}
 		})
-		database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-		database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(true, nil).Once()
 		codeIssuer.On("CreateAuthCode", issuanceTx, mock.Anything).
 			Return(nil, errors.WithStack(oauth.ErrIssuingClientGone)).Once()
@@ -893,8 +953,6 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 
 		boom := errors.New("the insert failed")
 		stub := expectRunInTransaction(database, issuanceTx)
-		database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-		database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(true, nil).Once()
 		codeIssuer.On("CreateAuthCode", issuanceTx, mock.Anything).Return(nil, boom).Once()
 		httpHelper.On("InternalServerError", rr, req, mock.MatchedBy(func(err error) bool {
@@ -938,8 +996,6 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 		stubLiveSession(database, 123)
 
 		stub := expectRunInTransaction(database, issuanceTx)
-		database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-		database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 		database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(false, nil).Once()
 		authHelper.On("ClearAuthContext", rr, req).Return(nil)
 
@@ -987,35 +1043,9 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 				},
 			},
 			{
-				name: "the user row acquisition fails",
-				setup: func(database *mocks_data.Database) *runInTransactionStub {
-					stub := expectRunInTransaction(database, issuanceTx)
-					database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(boom).Once()
-					// Neither row below is reached: a transaction that failed to take the top of
-					// the lock order has no business taking the next one down (#139 decision 11).
-					database.AssertNotCalled(t, "AcquireClientRowShared", mock.Anything, mock.Anything)
-					database.AssertNotCalled(t, "AcquireUserSessionRow", mock.Anything, mock.Anything)
-					return stub
-				},
-			},
-			{
-				name: "the client row acquisition fails",
-				setup: func(database *mocks_data.Database) *runInTransactionStub {
-					stub := expectRunInTransaction(database, issuanceTx)
-					database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-					database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(boom).Once()
-					// Same reasoning one row down: the session row is below clients in the order,
-					// so a transaction refused the client row does not go on to take it (#139).
-					database.AssertNotCalled(t, "AcquireUserSessionRow", mock.Anything, mock.Anything)
-					return stub
-				},
-			},
-			{
 				name: "the session row acquisition fails",
 				setup: func(database *mocks_data.Database) *runInTransactionStub {
 					stub := expectRunInTransaction(database, issuanceTx)
-					database.On("AcquireUserRow", issuanceTx, issuanceUserId).Return(nil).Once()
-					database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 					database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).
 						Return(false, boom).Once()
 					return stub
@@ -1070,96 +1100,6 @@ func TestHandleIssueGet_TheAcquisitionOrdersTheInsert(t *testing.T) {
 			})
 		}
 	})
-}
-
-// =============================================================================
-// #139 decision 11: the users row is taken before the user_sessions row
-//
-// codes.user_id is a foreign key, so the insert this transaction is about to make takes a lock on
-// the parent users row without ever naming it. Every credential operation, a password change, a
-// reset, an administrator setting a password, disabling or deleting an account, writes that row
-// first and reaches the sessions afterwards. Two orders, one cycle: measured on the branch that
-// introduced it, an authorization ceremony racing a password change deadlocks on MySQL and SQL
-// Server with the CREDENTIAL OPERATION as the victim, so the password does not change and the
-// session it was ending survives.
-//
-// The order is the whole of the remedy, so the order is what this pins. It cannot be seen at the
-// data tier, which measures the consequence on a real catalog but drives the statements by hand,
-// and it cannot be seen from the outside at all: both acquisitions are single-row UPDATEs that
-// leave nothing behind for an integration test to read. A mock recording the sequence is the only
-// place the handler's own choice is observable.
-// =============================================================================
-func TestHandleIssueGet_TheUserRowIsTakenBeforeTheSessionRow(t *testing.T) {
-	httpHelper := mocks_handlerhelpers.NewHttpHelper(t)
-	authHelper := mocks_handlerhelpers.NewAuthHelper(t)
-	templateFS := &mocks_test.TestFS{}
-	codeIssuer := mocks_oauth.NewCodeIssuer(t)
-	tokenIssuer := mocks_oauth.NewTokenIssuer(t)
-	database := mocks_data.NewDatabase(t)
-	auditLogger := mocks_audit.NewAuditLogger(t)
-	userSessionManager := mocks_user.NewUserSessionManager(t)
-	permissionChecker := mocks_user.NewPermissionChecker(t)
-
-	handler := HandleIssueGet(httpHelper, authHelper, templateFS, codeIssuer, tokenIssuer, database,
-		auditLogger, userSessionManager, permissionChecker)
-
-	req := requestWithSessionIdentifier(t, liveSessionIdentifier)
-	rr := httptest.NewRecorder()
-
-	authContext := &oauth.AuthContext{
-		AuthState:    oauth.AuthStateReadyToIssueCode,
-		Scope:        "openid profile",
-		ClientId:     "test-client",
-		UserId:       123,
-		ResponseMode: "query",
-		ResponseType: "code",
-		RedirectURI:  "https://example.com/callback",
-		State:        "test-state",
-	}
-	authHelper.On("GetAuthContext", req).Return(authContext, nil)
-	stubLiveSession(database, 123)
-
-	// Every statement on the transaction records itself, so the assertion below is about the
-	// sequence rather than about each call having happened somewhere.
-	var order []string
-	note := func(what string) func(mock.Arguments) {
-		return func(mock.Arguments) { order = append(order, what) }
-	}
-
-	// The stub notes the transaction's edges where the helper places them: "begin" before the
-	// body and "commit" after a nil return.
-	expectRunInTransaction(database, issuanceTx, func(edge string) { order = append(order, edge) })
-	database.On("AcquireUserRow", issuanceTx, issuanceUserId).Run(note("user row")).Return(nil).Once()
-	database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).
-		Run(note("client row")).Return(nil).Once()
-	database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).
-		Run(note("session row")).Return(true, nil).Once()
-	codeIssuer.On("CreateAuthCode", issuanceTx, mock.Anything).Run(note("insert")).
-		Return(&models.Code{Id: 1, Code: "test-code", ClientId: 1,
-			RedirectURI: "https://example.com/callback", State: "test-state"}, nil)
-
-	auditLogger.On("Log", constants.AuditCreatedAuthCode, mock.Anything).Return()
-	authHelper.On("ClearAuthContext", rr, req).Return(nil)
-
-	armIssueGate(database, userSessionManager, permissionChecker, authContext.RedirectURI)
-
-	handler.ServeHTTP(rr, req)
-
-	assert.Equal(t, []string{"begin", "user row", "client row", "session row", "insert", "commit"}, order,
-		"the lock order is users, then clients, then user_sessions, then the grants: the users row is "+
-			"the top because the insert reaches it again through the foreign key, and the shared "+
-			"client acquisition sits below it because a shared holder that afterwards reaches ABOVE "+
-			"clients closes a cycle with a client deletion queued exclusively on that row")
-
-	// And the ordering is only worth anything if all of it is on ONE transaction: a row taken on
-	// a different connection is released the moment that statement autocommits.
-	assert.Equal(t, http.StatusFound, rr.Code)
-	assert.Contains(t, rr.Header().Get("Location"), "code=test-code")
-
-	httpHelper.AssertExpectations(t)
-	authHelper.AssertExpectations(t)
-	database.AssertExpectations(t)
-	codeIssuer.AssertExpectations(t)
 }
 
 // =============================================================================
@@ -1559,20 +1499,6 @@ func TestHandleIssueGet_ImplicitAmbientSessionVanished(t *testing.T) {
 // parsing it.
 const liveSessionIdentifier = "session-identifier-abc"
 
-// issuanceUserId is the id armIssueGate's GetUserById hands back, and so the id the ceremony's
-// user-row acquisition names in every case that leaves that stub in place. Written out rather
-// than matched with mock.Anything: the point of the acquisition is that it takes the row of the
-// user this code is being minted for, and a stub matching any id would pass with the wrong one
-// (#139 decision 11). Cases that load a different user pass its id to
-// stubIssuanceTransactionFor.
-const issuanceUserId int64 = 1
-
-// issuanceClientId is the id armIssueGate's GetClientByClientIdentifier hands back, and so the id
-// the ceremony's shared client acquisition names. Written out for the reason issuanceUserId is:
-// the acquisition exists to take the row of the client this code is being minted for, and a stub
-// matching any id would pass with the wrong one (#139).
-const issuanceClientId int64 = 1
-
 // requestWithSessionIdentifier builds the /auth/issue request the way
 // MiddlewareSessionIdentifier leaves it when the session row exists: the identifier is in the
 // request context. Its absence is the terminated case, which is why the subtests that expect
@@ -1601,23 +1527,13 @@ func stubLiveSession(database *mocks_data.Database, ownerUserId int64) {
 var issuanceTx = &sql.Tx{}
 
 // stubIssuanceTransaction arms the transaction the authorization code branch opens around the
-// acquisition and the insert: the begin, an acquisition reporting the session still there, the
-// commit, and the deferred rollback, which runs on the success path too and is a no-op against a
-// committed transaction. A case that omits it fails on the strict mock.
+// acquisition and the insert: the helper, and an acquisition reporting the session still there.
+// A case that omits it fails on the strict mock.
 //
 // Cases about the refusal arm these themselves, because the answer they need from the acquisition
 // is the opposite one.
 func stubIssuanceTransaction(database *mocks_data.Database) {
-	stubIssuanceTransactionFor(database, issuanceUserId)
-}
-
-// stubIssuanceTransactionFor is the same, for a case whose ceremony loads a user other than the
-// one armIssueGate hands back. The id is a parameter rather than mock.Anything so that each case
-// states which row its ceremony takes (#139 decision 11).
-func stubIssuanceTransactionFor(database *mocks_data.Database, userId int64) {
 	expectRunInTransaction(database, issuanceTx)
-	database.On("AcquireUserRow", issuanceTx, userId).Return(nil).Once()
-	database.On("AcquireClientRowShared", issuanceTx, issuanceClientId).Return(nil).Once()
 	database.On("AcquireUserSessionRow", issuanceTx, liveSessionIdentifier).Return(true, nil).Once()
 }
 
@@ -4078,7 +3994,7 @@ func TestHandleIssueGet_RedirectURIRecheck(t *testing.T) {
 			if tc.wantIssued {
 				codeIssuer.On("CreateAuthCode", mock.Anything, mock.Anything).
 					Return(&models.Code{Id: 1, Code: "test-code", ClientId: 1, RedirectURI: tc.requested, State: "test-state"}, nil)
-				stubIssuanceTransactionFor(database, 123)
+				stubIssuanceTransaction(database)
 				auditLogger.On("Log", constants.AuditCreatedAuthCode, mock.Anything).Return()
 			} else {
 				auditLogger.On("Log", constants.AuditIssuanceRefusedRedirectURI, mock.MatchedBy(func(details map[string]interface{}) bool {
@@ -4413,7 +4329,7 @@ func TestHandleIssueGet_ScopeRefilter(t *testing.T) {
 						input.AuthContext.ConsentedScope == tc.wantConsented
 				})).Return(&models.Code{Id: 1, Code: "test-code", ClientId: 1,
 					RedirectURI: "https://example.com/callback", State: "test-state"}, nil)
-				stubIssuanceTransactionFor(database, 123)
+				stubIssuanceTransaction(database)
 				auditLogger.On("Log", constants.AuditCreatedAuthCode, mock.Anything).Return()
 			} else {
 				auditLogger.On("Log", constants.AuditIssuanceRefusedScopeDenied, mock.MatchedBy(func(details map[string]interface{}) bool {
