+++ src/core/data/mssqldb/user_session.go
@@ -107,10 +107,6 @@ func (d *MsSQLDatabase) AcquireUserSessionRow(tx *sql.Tx, sessionIdentifier stri
 	return d.CommonDB.AcquireUserSessionRow(tx, sessionIdentifier)
 }
 
-func (d *MsSQLDatabase) AcquireUserSessionRowById(tx *sql.Tx, userSessionId int64) error {
-	return d.CommonDB.AcquireUserSessionRowById(tx, userSessionId)
-}
-
 func (d *MsSQLDatabase) DeleteIdleSessions(tx *sql.Tx, idleTimeout time.Duration) error {
 	return d.CommonDB.DeleteIdleSessions(tx, idleTimeout)
 }
