+++ src/core/data/postgresdb/user.go
@@ -9,10 +9,6 @@ import (
 	"github.com/pkg/errors"
 )
 
-func (d *PostgresDatabase) AcquireUserRow(tx *sql.Tx, userId int64) error {
-	return d.CommonDB.AcquireUserRow(tx, userId)
-}
-
 func (d *PostgresDatabase) CreateUser(tx *sql.Tx, user *models.User) error {
 	now := time.Now().UTC()
