+++ src/core/data/sqlitedb/client.go
@@ -22,23 +22,6 @@ func (d *SQLiteDatabase) AcquireClientRow(tx *sql.Tx, clientId int64) error {
 	return d.CommonDB.AcquireClientRow(tx, clientId)
 }
 
-// AcquireClientRowShared is the EXCLUSIVE acquisition here, and the substitution costs nothing:
-// SQLite has one writer, so there is no shared mode for it to be shared with. The exclusive
-// acquisition is what the other three engines' shared statement conflicts with anyway, so this is
-// the same order written in the only lock this engine has.
-//
-// IT IS ALSO A WRITE ON PURPOSE, and a plain locking SELECT here would be a regression rather than
-// a cheaper equivalent. SQLite answers a transaction that read first and then tries to write with
-// SQLITE_BUSY at once, without consulting the busy handler, so a `SELECT` here would turn every
-// sign-in and every session bump into a reader upgrading to a writer: two of them overlapping
-// would fail immediately instead of waiting. Unreachable in one process, where
-// sqlitedb/db.go calls SetMaxOpenConns(1) and the two never overlap at all, and fail-closed
-// across processes, but there is no reason to buy it. Measured: with the SELECT spelling, the
-// data tier's discovery-barrier case fails here and passes everywhere else (#139).
-func (d *SQLiteDatabase) AcquireClientRowShared(tx *sql.Tx, clientId int64) error {
-	return d.CommonDB.AcquireClientRow(tx, clientId)
-}
-
 func (d *SQLiteDatabase) GetClientById(tx *sql.Tx, clientId int64) (*models.Client, error) {
 	return d.CommonDB.GetClientById(tx, clientId)
 }
