+++ src/core/user/usersession_manager_test.go
@@ -182,7 +182,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// Verify the session was updated with new ACR and AuthMethods
 			return s.AcrLevel == enums.AcrLevel2Optional.String() &&
@@ -214,7 +213,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			return s.AcrLevel == enums.AcrLevel2Mandatory.String() &&
 				s.AuthMethods == "pwd otp"
@@ -245,7 +243,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// ACR should upgrade, AuthMethods should stay the same
 			return s.AcrLevel == enums.AcrLevel2Mandatory.String() &&
@@ -276,7 +273,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// ACR should NOT be downgraded, should remain level2_mandatory
 			return s.AcrLevel == enums.AcrLevel2Mandatory.String()
@@ -307,7 +303,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			return s.AcrLevel == enums.AcrLevel2Optional.String()
 		})).Return(nil)
@@ -336,7 +331,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			return s.AcrLevel == enums.AcrLevel2Optional.String() &&
 				s.AuthMethods == "pwd otp"
@@ -365,7 +359,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// AuthMethods should remain "pwd" when empty string passed
 			return s.AuthMethods == "pwd"
@@ -396,7 +389,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// AcrLevel should remain level2_optional when empty string passed
 			return s.AcrLevel == enums.AcrLevel2Optional.String()
@@ -427,7 +419,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// Both should be preserved
 			return s.AcrLevel == enums.AcrLevel2Mandatory.String() &&
@@ -460,7 +451,6 @@ func TestBumpUserSession_StepUpAuthentication(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// AuthMethods should be updated
 			return s.AuthMethods == "pwd otp"
@@ -530,7 +520,6 @@ func TestBumpUserSession_ClientTracking(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, userSession).Return(nil)
 		database.On("UpdateUserSessionClient", mock.Anything, mock.Anything).Return(nil)
 		database.On("CreateUserSessionClient", mock.Anything, mock.MatchedBy(func(c *models.UserSessionClient) bool {
@@ -571,7 +560,6 @@ func TestBumpUserSession_ClientTracking(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, userSession).Return(nil)
 		database.On("UpdateUserSessionClient", mock.Anything, mock.MatchedBy(func(c *models.UserSessionClient) bool {
 			// LastAccessed should be updated to a newer time
@@ -609,7 +597,6 @@ func TestBumpUserSession_ClientTracking(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// IP should be concatenated
 			return s.IpAddress == "192.168.1.1,10.0.0.1"
@@ -646,7 +633,6 @@ func TestBumpUserSession_ClientTracking(t *testing.T) {
 		database.On("UserSessionLoadClients", mock.Anything, userSession).
 			Return(nil)
 		expectRunInTransaction(database, nil)
-		database.On("AcquireClientRowShared", mock.Anything, mock.Anything).Return(nil)
 		database.On("UpdateUserSession", mock.Anything, mock.MatchedBy(func(s *models.UserSession) bool {
 			// IP should NOT be duplicated
 			return s.IpAddress == "192.168.1.1"
